Moved file to features folder
