+++ src/features/add-time-machine-links-to-comments.js
@@ -1,8 +1,8 @@
 import select from 'select-dom';
 import {h} from 'dom-chef';
 
-import * as icons from './icons';
-import {getRepoURL} from './page-detect';
+import * as icons from '../libs/icons';
+import {getRepoURL} from '../libs/page-detect';
 
 export default async () => {
 	const comments = select.all('.timeline-comment-header:not(.rgh-timestamp-tree-link)');
